requirement.txt
numpy==1.26.4
torch==2.2.1
torchvision==0.17.1
Pillow==10.3.0
matplotlib==3.8.4
tqdm==4.66.4